Add SlackNotification
This is based on Porkchop's use of slack-notification. I've moved
the @notifier into the initialize to allow configuration with custom
options. This is put to use by the spec, which passes in a no-op
httpclient so that we don't try posting in specs.

The `return unless webhook_url` would have prevented posts just as well,
but wouldn't have given us much to assert against.

Change-Id: I6a5a362961584d817e4599513e7b4e9d2810c875