Add attachments to SlackNotification
Since we're going to need the rails url helpers in order to provide links
back to imperilment paths, so I think we might as well provide a basic
demo attachment here on the baseclass.

The goal here isn't to demonstate every possible attachment option, just
to show the basic form and include the "color" they will all include.

Change-Id: I39857ce972e568cffa9657069aa14a1ed24d0ba5