Don't show inactive answers as 'unanswered'
Icons for answers per user on the leaderboard are not currently
following the same rules as those on individual answers pages.
Specifically, inactive answers are shown as 'unanswered', instead of
'unavailable'.

This commit adjusts the logic to match that in
LeaderBoardHelper.status_for, accounting for the active status of an
answer.