Add date navigation buttons
Add two buttons which allow the user to jump one day ahead/behind, so
long as there is an answer for that day. Add two glyphs for the arrows
on the buttons and a helper method link_to_offset for linking to an
answer that is a certain number of days away.