Update controller request signatures
All of these updates boil down to two things

A) A change to the request helpers signature. Instead of params
as the second arg they're now a named option.
B) Implied params for nested resources that now must be stated
explicitly (eg: providing a game_id for a game/answer update)