Use Postgres in all environments
Sqlite is nice and fast but it also introduces inconsistencies.
For instance, it doesn't have a native BOOLEAN type, it uses 0
and 1. This is causing a problem for some of our queries against
the games.locked column. Instead of sorting that out I'd rather
just use PG everywhere and open the possibility of using PG
special features.