Replace FactoryGirl with FactoryBot
This gem has been renamed and this upgrade required some small
syntax changes to the factories (attributes now require a block).

There was one spec/controllers/questions_controller_spec.rb which
was passing but the context description did not match the setup.
It wasn't clear which combination the author was intended to
demonstrate so I just added a test to cover both cases.