Fix favicon on deployed environments
Previously the favicon was visible in dev because rails serves
the contents of /public automatically. In prod it will only
serve assets that are included in the precompile process. Here
I explicitly specify a favicon path by adding a link tag to the
layout, then move the icon to the default location.