Add separate classes for holding leaderboard results
Since we're currently building our leaderboard from the actual
models used to store the game info, we're limited in how we can
offer "perspective" on a game. In order to show something like
"a game_result as seen by this participant" we would need to
pass context down through all of the levels of building a result
so that we could filter associations.

By introducing a separate interface we can instead feed in only
the information that the current user should be privy to.