Adopt Point class for leader board result
In order to create a distinction between the Question as the
source of truth and what a user sees as the outcome of a game, we
introduce a class to hold their interpretation.

We also include the value for each question result so that we
will be able to calculate player totals on-the-fly.