Use question mark icon for unanswered questions
Hopefully this is more intuitive during on ongoing game. Now that
we display a mix of "unmarked" and "unavailable" for questions of
other players, we're able to be more leading with this icon. A
question mark makes a better CTA than an asterisk.

It looks a little odd for incomplete _locked_ game though, but
that's only because we haven't updated that logic. Now we begin
the harder shift from "how this question IS" to "what should you
DO". "unanswered" is a status, it's a fact. But a distinction can
be made between "unanswered because you haven't" and "unanswered
because you can't". We want to let players know how they can
interact with the board.