Choose Answer start_date based on game state
Instead of choosing sequential start_dates for all Answers, we
should suggest dates that make sense for the particular game.
We choose the dates which cover the week which precedes the
ended_at date.

The previous logic worked well for

A) new apps which would like the first answer to be dated today
B) apps with no gaps between weeks (so it was fine to look to the
last answer).

The new logic avoids surprising date suggestions for particular
answers.