Fix spec start_date collisions
The Answer factory uses a sequence for start_date that conflicts
with these values. A simple way to avoid the problem is to move
these hard coded values back to dates that won't conflict.

The problem was unoticed because none of these specs persist any
of the factory built Answers.