Fix flakey time distance spec
Math between Time and Date can get confusing. Here we're asking
for an approximation of the distance to "5 days in the future".
5.days.from_now will always advance the Time by exactly 120 hours
and give you back the same Date when that Time is truncated to the
start of the day. But when compared aginst the current Time you
could be a distance that rounds up to 5 or down to 4 days.

By picking a current Time past noon we ensure that we'll always be
on the lower side of that rouding to get the message this spec
was asserting against.

In the future I'd like to eliminate this ambiguity by moving away
from Date.