Allow multiple answer with same start date
Dropping the uniqueness constraint will allow for a more work-
friendly release schedule. Previously the answers were available
as soon as they were entered. The start_date did correspond to
the Calendar dates but didn't do much in the app. Now that I'm
using it to gate visibility, I'll need the ability to post
multiple questions on the same day in order to keep a Mon-Fri
game plan.