Fix breakdown bar funkiness for non-uniform start dates
One of the casualties of my migration away from "generalized
game model" towards something that is more tightly coupled to
a work week is this method. It was originally intended to return
an array with a size equal to length in days of the Game. By
measuring from the earliest answer start_date to the game's
ended_at it would guess at how many Answer you intended to input.
Then it would populate that array by walking the date range and
matching answer start_date.

In a world where all Games track a single week this boils down to
a really complicated way to count to seven. Here I cut to the
chase but introduce a subtle ordering problem. For now the order
can be inferred by the Answer index as long as they're entered
in the same order they appear in the calendar.