Add method to Game for calculating user max_wager
It is true that a player can definitely wager up to their current
score (all of the money they've earned from correct responses).
But it might turn out that an even higher wager is possible if
there remain unmarked questions. This method includes those amounts
by summing the amounts for all non-final, not-wrong (missing,
correct, or unmarked) questions.

This should be helpful for players who would like to confidently
place a large wager and proceed with final imperilment before
waiting for an admin to mark their responses.