Re-check final wager when updating any question
This is the controller action that handles question marking.
Since we're now allowing users to overbid on final imperilment,
we need to revue that wager anytime any question for the same
game could transition to or from "incorrect".

We could use a model save callback to be more robust, but I'd like
to see how long we can get away with an explicit call in a
controller.