Fix bug where max_wager combined all players
I forgot to limit the wager calculation to the user in question.
That mistake wasn't caught by the test because these were no
questions belonging to other users.

Unfortunately in order to get that constraint into the JOIN clause
we need to write some fancy arel. This ensures that we only match
answers with a question that belongs to this answer AND the user
of interest. We still require that the questions be correct or
unmarked. In our case `correct = NULL` is serving double-duty.
It's matching both joined rows that have a question which has not
been marked AND rows that don't have a match because this user
has not responded.