Fix error loading questions
Although normally permissions are additive, this `can :manage` is
actually causing problems for the `can :index, Question` lower
down. When performing load_and_authorize_resource without an
instance of a resource (eg: a collection route like :index) there
is a line of code that checks to see if ANY of the permissions
have a block. In that situation it short circuits and refuses
to load the collection.

By splitting this up from a :manage to the many sub-actions (even
though most are probably not actually used) I can preserve the
block logic.

It would probably be a good time to review all of the permissions.
Complex ability files can have counter-intuitive interactions and
the specs don't do much to explain our intentions. Because we're
using custom matchers to assert against the rules themselves,
we've basically just copied the Ability file in a slightly
different syntax.