Fix question marking "undo"
These links communicate with the backend via AJAX. We use the
markup to store the 3 potential values of this attribute (true
for correct, false for incorrect, and nil for unmarked).

In Rails 5 they started stripped `nil` values from nested tag
options (eg: the type of hash notation we used here to avoid
the top level pruning of `nil` value options). By explicitly
passing the string 'null' instead I can simulate the old
behaviour. The backend is already looking for this string in
questions#update.