Do not require password confirmation to update user
Google authed users don't know their password since we generate
it for them. It's also not easy to distinguish those accounts
from those that use the login form.

Although it is possible to reset your password, that's not the
most intuitive action when confronted with this user edit page
for the first time. For now, I would like to drop password
confirmation to ease adtoption of the new display_name.