Add display_name hint to leaderboard
To encourage adoption of a display_name without being overly
annoying, I would like to show an asterisk next to the current
user's entry in the leaderboard if they have not chosen a
display_name. This asterisk has a tooltip and is a link to the
user edit page where they can pick a display_name.