End all games on a Sunday
Our game shedule is centered around the cadence of a 5 day work
week. Currently the game#new form will force you to pick a
Sunday as the ended_at date, with JS. This change codifies that
restriction so that we can get fancy will the schedule for
rolling out the questions throughout the week.